revise percent to be integer

+++ db/migrate/20230422205757_create_discounts.rb
@@ -2,7 +2,7 @@ class CreateDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :discounts do |t|
       t.references :merchant, foreign_key: true
-      t.float :percent
+      t.integer :percent
       t.integer :quantity
     end
   end
